+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix for energy data that was not being stored for the last job of the batch #2657 #2418
 - Reduce the frequency with which energy data was not being stored for the last job of the batch #2657 #2418
 - Fixed issue with experiments running on LOCAL platform without a defined LOCAL entry in platforms.yml config file #1131
+- Fixed an issue with the ECMWF platform not working with the new online recovery
 
 **Enhancements:**